+++ frontend/src/app/exames/components/material-exames/material-exames.component.css
@@ -14,3 +14,14 @@
 .label-radioButton {
     margin: 5px;
 }
+
+.inputNumber {
+    display: flex;
+    flex-direction: column;
+}
+
+.label {
+    display: flex;
+    overflow: hidden;
+    text-overflow: ellipsis;
+}
